backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langchain-text-splitters==0.3.4
chromadb==0.6.3
cohere==5.13.3
groq==0.15.0
pypdf==5.1.0
python-frontmatter==1.1.0
markdown==3.7
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
respx==0.22.0
ruff==0.8.6
black==24.10.0
